chromeos-installer: avoid unnecessary mtdblock flush

+++ chromeos-base/chromeos-installer/files/flash_spi.sh
@@ -13,9 +13,14 @@ main() {
     loader="${LOADER_DIR}/rkspi_loader.img"
  fi
 
- echo "installing $loader to $target"
- dd if=$loader of=$target status=progress bs=1024 conv=fdatasync
- echo "done"
+ local target_md5="$(md5sum ${target} | awk '{print $1}')"
+ local loader_md5="$(md5sum ${loader} | awk '{print $1}')"
+
+ if [ "$target_md5" != "$loader_md5" ]; then
+    echo "installing $loader to $target"
+    dd if=$loader of=$target status=progress bs=1024 conv=fdatasync
+    echo "done"
+ fi
 }
 
 main "$@"
